requirements.txt
# Document Processing Pipeline Dependencies
dropbox==11.36.2
pinecone>=5.0.0
openai>=1.97.0
pdfplumber==0.9.0
pandas>=2.1.4
numpy>=1.26.0,<2.0  # Pin to 1.x for Docling ABI compatibility
python-dotenv==1.0.0
openpyxl>=3.1.5
xlrd==2.0.1
python-docx>=1.1.2
python-pptx>=0.6.21             # PowerPoint (.pptx) file parsing
pdf2image==1.16.3
Pillow>=10.0.0               # Image processing
langchain-experimental       # For SemanticChunker
# sentence-transformers removed - using Pinecone embedding models
# scikit-learn removed - using Pinecone embedding models
click==8.1.7
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.9.0
celery==5.3.4
redis==5.0.1
boto3==1.34.0
pytest==7.4.3
tenacity==8.2.3
colorama==0.4.6
extract-msg>=0.46.0
docx2txt>=0.8
filelock>=3.12.0
docling>=2.54.0              # Granite Docling document converter (PDF + OCR + tables)
transformers>=4.56.2         # Required by Docling for VLM/image-text models
torch>=2.0.0                 # Required by transformers/Docling (CPU build is sufficient